Updated dependencies, rewrote a couple settings screens, removed button to send bug report, removed telegram & website from about screen

+++ app/src/main/java/posidon/launcher/feed/notifications/NotificationItem.kt
@@ -49,7 +49,7 @@ class NotificationItem(
         result = 31 * result + (actions?.contentHashCode() ?: 0)
         result = 31 * result + (contentIntent?.hashCode() ?: 0)
         result = 31 * result + key.hashCode()
-        result = 31 * result + progress.toInt()
+        result = 31 * result + progress
         return result
     }
 }
